% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hysplit_trajectory.R
\name{download_met_files}
\alias{download_met_files}
\title{Download Meteorology Files}
\usage{
download_met_files(
  days,
  duration,
  direction,
  met_dir = paste0(getwd(), "/meteorology")
)
}
\arguments{
\item{days}{Dates for which to pull meteorology data}

\item{duration}{Duration of desired HYSPLIT run; determines how many dates
before/after "days" are pulled}

\item{direction}{Direction of desired HYSPLIT run; determines whether to pull
additional days before or after "days"}

\item{met_dir}{Path to directory where meteorology files are to be written}
}
\description{
Download Meteorology Files
}
