% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hysplit_trajectory.R
\name{hysplit_trajectory}
\alias{hysplit_trajectory}
\title{Calculate HYSPLIT Trajectories}
\usage{
hysplit_trajectory(
  run_df = NULL,
  lat = NULL,
  lon = NULL,
  height = 500,
  duration = 9,
  days = NULL,
  daily_hours = 0,
  direction = "forward",
  vert_motion = 0,
  model_height = 20000,
  extended_met = TRUE,
  vbug = 2.5,
  traj_name = NULL,
  binary_path1 = NULL,
  met_dir = paste0(getwd(), "/meteorology"),
  exec_dir = NULL,
  clean_up = TRUE,
  local_time = FALSE
)
}
\arguments{
\item{run_df}{A data frame whose rows contain the information for each
trajectory run. This data frame should contain the following columns: "lat",
"lon", "height", "date" and "hour"}

\item{lat}{Latitude}

\item{lon}{Longitude}

\item{height}{Dispersal altitude}

\item{duration}{Number of hours for the trajectories to run}

\item{days}{Dates for the trajectory runs (as character or Date)}

\item{daily_hours}{Starting hours for the trajectory (numeric)}

\item{direction}{One of "forward" or "backward", denoting the direction of
the HYSPLIT run}

\item{extended_met}{TRUE or FALSE; should extra meteorological columns be
included in the output?}

\item{vbug}{Moth velocity}

\item{traj_name}{Name of output folder}

\item{met_dir}{Path to directory containing downloaded meteorological data}

\item{exec_dir}{Location of main directory}

\item{clean_up}{TRUE or FALSE; delete all run folders on completion?}

\item{local_time}{TRUE or FALSE; the input times are local relative to the
input latitude and longitude. If FALSE, it is assumed that times are in GMT.}
}
\description{
Calculate HYSPLIT Trajectories
}
